Moved the function to represent an APE 14-compliant as a string to a standalone function and use only in SlicedLowLevelWCS, not BaseLowLevelWCS
