+++ astropy/wcs/wcsapi/__init__.py
@@ -1,4 +1,5 @@
 from .low_level_api import *
 from .high_level_api import *
 from .high_level_wcs_wrapper import *
+from .utils import *
 from .sliced_low_level_wcs import *
