+++ astropy/wcs/wcsapi/sliced_low_level_wcs.py
@@ -1,6 +1,6 @@
 import numbers
 import numpy as np
-from astropy.wcs.wcsapi import BaseLowLevelWCS
+from astropy.wcs.wcsapi import BaseLowLevelWCS, wcs_as_str
 
 __all__ = ['sanitize_slices', 'SlicedLowLevelWCS']
 
@@ -213,4 +213,10 @@ class SlicedLowLevelWCS(BaseLowLevelWCS):
 
     @property
     def axis_correlation_matrix(self):
-        return self._wcs.axis_correlation_matrix[self._world_keep][:,self._pixel_keep]
+        return self._wcs.axis_correlation_matrix[self._world_keep][:, self._pixel_keep]
+
+    def __repr__(self):
+        return wcs_as_str(self)
+
+    def __str__(self):
+        return wcs_as_str(self)
