+++ astropy/wcs/wcsapi/utils.py
@@ -1,4 +1,8 @@
 import importlib
+import numpy as np
+
+
+__all__ = ['deserialize_class', 'wcs_as_str']
 
 
 def deserialize_class(tpl, construct=True):
@@ -21,3 +25,86 @@ def deserialize_class(tpl, construct=True):
         return klass(*args, **kwargs)
     else:
         return klass, args, kwargs
+
+
+def wcs_as_str(wcs):
+
+    # Overall header
+
+    s = f'{wcs.__class__.__name__} Transformation\n\n'
+    s += ('This transformation has {} pixel and {} world dimensions\n\n'
+            .format(wcs.pixel_n_dim, wcs.world_n_dim))
+    s += f'Array shape (Numpy order): {wcs.array_shape}\n\n'
+
+    # Pixel dimensions table
+
+    array_shape = wcs.array_shape or (0,)
+    pixel_shape = wcs.pixel_shape or (None,) * wcs.pixel_n_dim
+
+    # Find largest between header size and value length
+    pixel_dim_width = max(9, len(str(wcs.pixel_n_dim)))
+    pixel_siz_width = max(9, len(str(max(array_shape))))
+
+    s += (('{0:' + str(pixel_dim_width) + 's}').format('Pixel Dim') + '  ' +
+            ('{0:' + str(pixel_siz_width) + 's}').format('Data size') + '  ' +
+            'Bounds\n')
+
+    for ipix in range(wcs.pixel_n_dim):
+        s += (('{0:' + str(pixel_dim_width) + 'd}').format(ipix) + '  ' +
+                (" "*5 + str(None) if pixel_shape[ipix] is None else
+                ('{0:' + str(pixel_siz_width) + 'd}').format(pixel_shape[ipix])) + '  ' +
+                '{:s}'.format(str(None if wcs.pixel_bounds is None else wcs.pixel_bounds[ipix]) + '\n'))
+
+    s += '\n'
+
+    # World dimensions table
+
+    # Find largest between header size and value length
+    world_dim_width = max(9, len(str(wcs.world_n_dim)))
+    world_typ_width = max(13, max(len(x) if x is not None else 0 for x in wcs.world_axis_physical_types))
+
+    s += (('{0:' + str(world_dim_width) + 's}').format('World Dim') + '  ' +
+            ('{0:' + str(world_typ_width) + 's}').format('Physical Type') + '  ' +
+            'Units\n')
+
+    for iwrl in range(wcs.world_n_dim):
+
+        if wcs.world_axis_physical_types[iwrl] is not None:
+            s += (('{0:' + str(world_dim_width) + 'd}').format(iwrl) + '  ' +
+                    ('{0:' + str(world_typ_width) + 's}').format(wcs.world_axis_physical_types[iwrl]) + '  ' +
+                    '{:s}'.format(wcs.world_axis_units[iwrl] + '\n'))
+        else:
+            s += (('{0:' + str(world_dim_width) + 'd}').format(iwrl) + '  ' +
+                    ('{0:' + str(world_typ_width) + 's}').format('None') + '  ' +
+                    '{:s}'.format('unknown' + '\n'))
+    s += '\n'
+
+    # Axis correlation matrix
+
+    pixel_dim_width = max(3, len(str(wcs.world_n_dim)))
+
+    s += 'Correlation between pixel and world axes:\n\n'
+
+    s += (' ' * world_dim_width + '  ' +
+            ('{0:^' + str(wcs.pixel_n_dim * 5 - 2) + 's}').format('Pixel Dim') +
+            '\n')
+
+    s += (('{0:' + str(world_dim_width) + 's}').format('World Dim') +
+            ''.join(['  ' + ('{0:' + str(pixel_dim_width) + 'd}').format(ipix)
+                    for ipix in range(wcs.pixel_n_dim)]) +
+            '\n')
+
+    matrix = wcs.axis_correlation_matrix
+    matrix_str = np.empty(matrix.shape, dtype='U3')
+    matrix_str[matrix] = 'yes'
+    matrix_str[~matrix] = 'no'
+
+    for iwrl in range(wcs.world_n_dim):
+        s += (('{0:' + str(world_dim_width) + 'd}').format(iwrl) +
+                ''.join(['  ' + ('{0:>' + str(pixel_dim_width) + 's}').format(matrix_str[iwrl, ipix])
+                        for ipix in range(wcs.pixel_n_dim)]) +
+                '\n')
+
+    # Make sure we get rid of the extra whitespace at the end of some lines
+    return '\n'.join([l.rstrip() for l in s.splitlines()])
+
